@@ -47,23 +47,23 @@
     "e2e": "ng e2e"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.803.25",
-    "@angular-devkit/build-ng-packagr": "~0.803.25",
-    "@angular/cli": "~8.3.25",
-    "@angular/common": "^8.2.14",
-    "@angular/compiler": "^8.2.14",
-    "@angular/compiler-cli": "^8.2.14",
-    "@angular/core": "^8.2.14",
-    "@angular/forms": "^8.2.14",
-    "@angular/language-service": "^8.2.14",
-    "@angular/platform-browser": "^8.2.14",
-    "@angular/platform-browser-dynamic": "^8.2.14",
-    "@angular/router": "^8.2.14",
+    "@angular-devkit/build-angular": "~0.900.1",
+    "@angular-devkit/build-ng-packagr": "~0.900.1",
+    "@angular/cli": "~9.0.1",
+    "@angular/common": "^9.0.0",
+    "@angular/compiler": "^9.0.0",
+    "@angular/compiler-cli": "^9.0.0",
+    "@angular/core": "^9.0.0",
+    "@angular/forms": "^9.0.0",
+    "@angular/language-service": "^9.0.0",
+    "@angular/platform-browser": "^9.0.0",
+    "@angular/platform-browser-dynamic": "^9.0.0",
+    "@angular/router": "^9.0.0",
     "@earshinov/extract-scss-variables": "^0.1.1",
     "@types/jasmine": "~3.4.0",
     "@types/jasminewd2": "~2.0.6",
-    "@types/node": "~12.7.1",
-    "codelyzer": "^5.1.0",
+    "@types/node": "^12.11.1",
+    "codelyzer": "^5.1.2",
     "jasmine-core": "~3.4.0",
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "~4.2.0",
@@ -74,17 +74,18 @@
     "karma-jasmine": "~2.0.1",
     "karma-jasmine-html-reporter": "^1.4.2",
     "karma-safari-launcher": "^1.0.0",
-    "ng-packagr": "^5.7.1",
+    "ng-packagr": "^9.0.0",
     "node-sass": "^4.12.0",
     "protractor": "~5.4.0",
     "rxjs": "~6.5.4",
     "scss-bundle": "^2.5.1",
     "ts-node": "~8.3.0",
-    "tsickle": "^0.36.0",
-    "tslib": "^1.10.0",
     "tslint": "~5.18.0",
-    "typescript": "~3.5.3",
-    "zone.js": "~0.9.1"
+    "typescript": "~3.7.5",
+    "zone.js": "~0.10.2"
   },
-  "peerDependencies": {}
+  "peerDependencies": {},
+  "dependencies": {
+    "tslib": "^1.10.0"
+  }
 }
