@@ -31,7 +31,7 @@ import {checkWizardState} from '../util/test-utils';
   `
 })
 class WizardTestComponent {
-  @ViewChild(WizardComponent, { static: false })
+  @ViewChild(WizardComponent)
   public wizard: WizardComponent;
 }
 
